+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace Goodcat\L10n\Tests;
 
 use Goodcat\L10n\L10nServiceProvider;
+use Laravel\Wayfinder\WayfinderServiceProvider;
 
 class TestCase extends \Orchestra\Testbench\TestCase
 {
@@ -10,6 +11,7 @@ class TestCase extends \Orchestra\Testbench\TestCase
     {
         return [
             L10nServiceProvider::class,
+            WayfinderServiceProvider::class,
         ];
     }
 }